Infra/refine setups (#3)
* refine

* refine

* fix auto

* move port to env

* refine

* refine

* refine

* fix env

* finish

+++ scripts/infra_destroy.sh
@@ -2,6 +2,13 @@
 
 set -e
 
+# Check and rn in the infra folder
+if [ ! -d "infra" ]; then
+  echo "Error: 'infra' folder not found. Exiting..."
+  exit 1
+fi
+cd infra
+
 # Destroy Terraform resources
 echo "1. Destroying Terraform resources..."
 terraform destroy -auto-approve
@@ -11,8 +18,7 @@ echo "2. Removing local Terraform files..."
 rm -rf \
   .terraform* \
   terraform.tfstate* \
-  tfplan \
-  generated-task-definition.json
+  tfplan
 
 # Delete AWS secrets
 echo "3. Cleaning up AWS secrets..."
